Refactored project to be called "Configurable slayer task overlay" rather than "Configurable slayer info"
