+++ runelite-plugin.properties
@@ -1,6 +1,6 @@
-displayName=Configurable Slayer Info
+displayName=Configurable Slayer Task Overlay
 author=mathiaslj, BrastaSauce, wesley221
 support=
 description=Configurable information to display in overlay for slayer tasks
 tags=slayer,overlay,task,configurable
-plugins=com.mathiaslj.configurableslayerinfo.ConfigurableSlayerInfoPlugin
+plugins=com.mathiaslj.configurableslayertaskoverlay.ConfigurableSlayerTaskOverlayPlugin
