+++ src/main/java/com/mathiaslj/configurableslayertaskoverlay/models/NpcLocation.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package com.mathiaslj.configurableslayerinfo.models;
+package com.mathiaslj.configurableslayertaskoverlay.models;
 
 import lombok.Getter;
 import net.runelite.api.coords.WorldArea;
