+++ src/test/java/com/mathiaslj/configurableslayertaskoverlay/ConfigurableSlayerTaskOverlayTest.java
@@ -1,11 +1,11 @@
-package com.mathiaslj.configurableslayerinfo;
+package com.mathiaslj.configurableslayertaskoverlay;
 
 import net.runelite.client.RuneLite;
 import net.runelite.client.externalplugins.ExternalPluginManager;
 
-public class ConfigurableSlayerInfoTest {
+public class ConfigurableSlayerTaskOverlayTest {
     public static void main(String[] args) throws Exception {
-        ExternalPluginManager.loadBuiltin(ConfigurableSlayerInfoPlugin.class);
+        ExternalPluginManager.loadBuiltin(ConfigurableSlayerTaskOverlayPlugin.class);
         RuneLite.main(args);
     }
 }
